notifications are merged to the same place
