+++ alv-portal-ui/src/app/app.component.html
@@ -1,5 +1,6 @@
 <alv-header></alv-header>
 
+
 <div class="sr-only"
      role="status"
      aria-live="polite"
