+++ alv-portal-ui/src/app/home/pages/job-seeker-home-page/job-seeker-home-page.component.html
@@ -22,4 +22,6 @@
       <button class="btn btn-primary">Show job offers</button>
     </div>
   </div>
+  <alv-notifications-container></alv-notifications-container>
+
 </alv-panel>
