--- alv-portal-ui/src/app/shared/layout/notifications/message-type.enum.ts
@@ -1,7 +1,0 @@
-export enum MessageType {
-  SUCCESS = 'success',
-  ERROR = 'error',
-  INFO = 'info',
-  WARNING = 'warning',
-  EMPTY = 'empty'
-}
