+++ alv-portal-ui/src/app/shared/layout/notifications/notification.model.ts
@@ -1,0 +1,14 @@
+export interface Notification {
+  id?: number;
+  type: NotificationType;
+  messageKey: string;
+  isSticky: boolean;
+}
+
+export enum NotificationType {
+  DEFAULT,
+  INFO,
+  SUCCESS,
+  WARNING,
+  ERROR
+}
