+++ alv-portal-ui/src/app/shared/layout/notifications/notification/notification.component.html
@@ -1,10 +1,7 @@
 <div class="icon">
-  <i [class]="getIconClass(type)"></i>
+  <i [class]="getIconClass(notification.type)"></i>
 </div>
 <div class="text">
-  <div class="line-1 text-truncate"> {{ message }}</div>
-  <div class="line-2 text-truncate"
-       *ngIf="additionalText">{{ additionalText }}
-  </div>
+  <div class="line-1 text-truncate"> {{ notification.messageKey }}</div>
 </div>
