+++ alv-portal-ui/src/app/shared/layout/notifications/notifications-container/notifications-container.component.html
@@ -1,1 +1,6 @@
+<div *ngFor="let notification of notifications">
 
+  <alv-notification [notification]=""></alv-notification>
+
+
+</div>
